Dev (#23)
* licence.AesService->utils.AesUtil

* 手动安装本地依赖方法记录

* 通过定义MyPage实现自定义分页

* 自定义注解添加注释

* 添加HttpResult,添加统一异常处理和自定义异常，当抛出自定义异常时，可以组装一个异常HttpResult序列化后返回给前端

* 修改bug

* 复习

* 升级版本

* 整理依赖，添加外置dependency management模块——mydep
修改部分包名

* demo4ys

* 为文章查询接口添加redis缓存功能，可以不经过mysql而使用redis缓存来进行接口数据返回
修改yaml文件的引入nacos配置文件的方式
修改redis连接池为lettuce

* 添加两个模块

* 两个模块实现rabbitmq的发布及监听动作

* 测试错误的tomcat参数

* 两个模块实现rabbitmq的发布及监听动作

* sth changes

* some change

* 代码梳理

* MybatisPlus版本升级

* feat: 将旧的用户权限系统替换为新的用户权限系统 step1 替换用户表

* feat: 修改README及结构文档

* feat: 登陆时省略PersonVo的中转过程

* 添加接口文档-User

* <servlet-api.version>3.0-alpha-1</servlet-api.version>

* user restful

* feat: 添加工具类和transformer

* feat: 去掉无用对象

* feat: 完成user相关的接口文档和接口

* feat: 修改MybatisPlus配置以及部分User相关的Mapper

* feat: 修改DataResult的默认成功失败方法的返回值

* feat: 添加UserInfo的管理

* feat: DataResult添加getBuilder

* feat: UserInfo添加更新接口

* feat: 修改Mapper的容器内注册方式

* fix: DataResult内存溢出问题解决

* feat: 自定义错误原因的DataResult

* feat: DataResult添加可以自定义原因的success方法

* feat: 添加userInfo的接口文档和对应的方法实现

* feat: 修改userInfo的接口

* feat: 改造HttpResult

* feat: 添加role相关的接口 修改之前的DataResult的用法

* refactor: 改造MyPage

* refactor: 修改DataResult和接口文档

* feat: 添加角色相关接口，增强MyPage的构造器的兼容性

* feat: 添加Access相关的rest接口

* refactor: 统一异常处理修改 MyPage重写toString

* feat: 添加Access的添加接口

* feat: 添加Group分页获取接口

* refactor: 修改数据库表名

* feat: 添加Group创建接口

* feat: Access添加 根据id获取、根据id删除、根据id更新的接口

* refactor: Transformer添加入参为空的处理

* feat: access添加日志并修改部分异常信息

* feat: group 添加根据id查询的接口

* feat: group 添加根据id删除、根据id更新的接口

* feat: role service层添加方法

* feat: 添加通用的逻辑删除mapper接口

* feat: 添加role的 根据id删除、根据id修改、根据id获取等接口

* feat: 添加user和role的上层service UserRoleService 并添加根据userId获取用户的角色列表的接口

* feat: role 添加根据roleId获取权限列表的接口

* fix: xml错误修正

* feat: 添加UserAccessMapper 可以直接获取用户的所有权限

* feat: 清理未使用的import

* feat: 重写各实体类及其DTO VO的toString方法

* feat: isWindows改为静态方法

* feat: 修改拦截器逻辑、返回信息

* refactor: 重写拦截器的权限校验逻辑 结合RBAC

* feat: 完成拦截器中的鉴权动作，当PermissionChecker没有实现时自动使用默认实现完成动作

* feat: 添加userType属性 区分用户的类型并作为hasPermission默认实现方式

* feat: 完成AdminPermission校验与角色权限的绑定 支持权限的路径中使用**代替任意字符串（支持/user/**, 不支持/user/**/hello）

* feat: 需要同时满足
      1. 接口所需的权限 >= 用户的userType
      2. 当前用户的权限列表中存在该url
      才可以访问该接口

* feat: 动态修改权限校验的级别 1 校验type 2 校验type和url

* feat: 修改默认bean的包名（默认bean 当容器中不存在时自动创建的bean）

* feat: 清理部分module

* feat: 添加依赖以及群管理的相关类

* refactor: 修改模块名称 ds->business

* feat: 添加talk模块 添加部分基础依赖

* feat: 整理pom版本号01

* feat: maven3不支持version为变量 将版本号统一为1.0.0-SNAPSHOT

* feat: 记录版本号修改命令，修改版本号为1.0.1-SNAPSHOT

* feat: 修改版本号为1.2.0-SNAPSHOT

* feat: 端口号统一在父pom中管理 使batch服务可用

* feat: 开发talk模块

* feat: 打通MQ

* feat: 添加 父菜单id、请求类型 字段

* feat: 添加websocket功能 待添加websocket的鉴权

* feat: 单测

* feat: WebsocketSession 添加属性成功

* feat: model-util 添加常量

* feat: 统一MQ常量

* feat: 统一MQ常量

* feat: 集成Redisson 支持分布式锁

* feat: blog集成Redisson

* feat: pom依赖冲突

* feat: log

* feat: 添加Files模块

* feat: 清理依赖

* feat: 完成文件上传接口并在最后删除临时文件

* feat: 清理依赖 解决chat的大部分hbase带来的冲突

* feat: 搭建Mybatis环境 FileUpload表

* feat: 完成文件上传和上传信息的入库功能

* feat: MyLogger代替Logger

* feat: 去掉无用参数

* feat: 修改pom 修改接口返回参数

* feat: 七牛云添加文件查询功能

* feat: 处理BUG

* feat: logicer-base已重命名类名

* feat: utils-server支持暴露服务的health监控端口

* feat: 解决引入的jackson与actuator之间的依赖冲突

* feat: AesUtil修改并添加单测

* feat: 解析七牛云返回信息时减少打印日志

* feat: 待解决自定义sql语句

* feat: 仍然待解决自定义sql语句

* feat: 添加ES

* feat: 日志配置logback

* feat: 自定义logback的MessageConverter 在打印json格式的日志到文件时，将日志字符串双引号前添加一个转义符 以免json解析失败

* feat: json文件写入时每条json占一行

* feat: json文件写入时每条json占一行 每条json结尾都带换行符

* feat: message属性名称改为strmsg

* feat: 记一次坑

* fix: 空指针

* fix: 统一五个业务服务的日志输出规范

* fix: 去除Configuration注解 gateway单独添加Converter

* fix: json文件中将所有换行符\r\n改为空格

* fix: 修改article入库sql

* fix: 完善非七牛云方式的文件上传 文件存储到服务器本地

* fix: 文本文件与通用文件区分开 blog包含一套单独的文件上传，只用作文章文件上传 cm-file包含通用文件上传

* feat: 修改鉴权规则

* feat: 集成netty客户端并打通ws和netty服务

* feat: 完成前端与服务端的websocket交互

* feat: thread

* feat: 解决线程不释放引起的登录次数限制问题

* feat: netty断开连接时 ws也断开

* feat: 修改地址为线上

* feat: 添加HbaseService管理Hbase数据库

* feat: 连接参数修改

* feat: 添加Hbase的连接和管理功能

* feat: 修改test

* feat: 修改jar包生成位置

* feat: 文件转换接口改动

---------

+++ common/licence/src/main/java/com/mylog/common/licence/service/impl/UserServiceImpl.java
@@ -1,32 +1,35 @@
 package com.mylog.common.licence.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.baomidou.mybatisplus.core.conditions.update.UpdateWrapper;
-import com.baomidou.mybatisplus.core.metadata.IPage;
-import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
-import com.mylog.common.licence.enumcenter.GroupEnum;
+import com.baomidou.mybatisplus.core.toolkit.Wrappers;
+import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.mylog.common.licence.entity.User;
+import com.mylog.common.licence.enumcenter.GroupEnum;
 import com.mylog.common.licence.mapper.UserMapper;
 import com.mylog.common.licence.model.dto.UserDTO;
 import com.mylog.common.licence.model.vo.UserVO;
 import com.mylog.common.licence.service.IUserService;
-import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import com.mylog.common.licence.service.PasswordService;
-import com.mylog.entitys.entitys.entity.Message;
-import com.mylog.entitys.entitys.entity.Person;
-import com.mylog.entitys.entitys.entity.Result;
-import com.mylog.entitys.entitys.entity.Status;
-import com.mylog.entitys.entitys.vo.PersonVo;
+import com.mylog.common.licence.transformer.UserTransformer;
+import com.mylog.tools.model.model.entity.Person;
+import com.mylog.tools.model.model.info.Message;
+import com.mylog.tools.model.model.info.Status;
+import com.mylog.tools.model.model.page.MyPage;
+import com.mylog.tools.model.model.result.DataResult;
+import com.mylog.tools.model.model.result.HttpResult;
+import com.mylog.tools.model.model.result.PageDataResult;
+import com.mylog.tools.model.model.vo.PersonVo;
 import com.mylog.tools.utils.session.UserContext;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.mylog.tools.utils.utils.PasswordService;
+import com.mylog.tools.utils.utils.Safes;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.stereotype.Service;
 
+import javax.annotation.Resource;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 /**
  * <p>
@@ -35,17 +38,15 @@ import java.util.List;
  * @author Dylan
  * @since 2020-05-24
  */
+@Slf4j
 @Service
 @RefreshScope
 public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IUserService {
 
-    private static final Logger logger = LoggerFactory.getLogger(UserServiceImpl.class);
-
-    @Autowired
+    @Resource
     private UserMapper userMapper;
-    @Autowired
+    @Resource
     private PasswordService passwordService;
-
 
     /**
      * 获取用户列表
@@ -53,18 +54,20 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IU
      * @return
      */
     @Override
-    public Result selectUserList(Page<User> page) {
-        Result result = null;
-        IPage<User> userPage = userMapper.selectUserList(page);
-        List<User> userList = userPage.getRecords();
+    public HttpResult selectUserList(MyPage page) {
+        List<User> userList = userMapper.selectUserList(page);
         List<UserVO> userVOList = new ArrayList<>();
-        for (User user:userList){
-            UserVO userVO = new UserVO();
-            BeanUtils.copyProperties(user, userVO);
-            userVOList.add(userVO);
-        }
-        result = new Result.Builder(Status.SUCCESS.getStatus(), Message.SUCCESS.getMsg()).data(userVOList).page(page.getCurrent()).size(page.getSize()).total(page.getTotal()).build();
-        return result;
+        Safes.of(userList).forEach(u -> {
+            userVOList.add(UserTransformer.user2UserVo(u));
+        });
+        log.info("MyPage : {}", page);
+        return PageDataResult
+                .success()
+                .data(userVOList)
+                .page(page.getPageNo())
+                .size(page.getPageSize())
+                .total(userMapper.selectUserTotal())
+                .build();
     }
 
     /**
@@ -73,32 +76,28 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IU
      * @return
      */
     @Override
-    public Result selectOne(UserDTO userDTO) {
-        Result result;
+    public HttpResult selectOne(UserDTO userDTO) {
+        DataResult dataResult;
         QueryWrapper<User> queryWrapper = new QueryWrapper<>();
         if (userDTO.getId()!=null && userDTO.getId()!=0){
             queryWrapper.eq("id", userDTO.getId());
         }
-        if (userDTO.getUsername()!=null&&userDTO.getUsername().length()>0){
-            queryWrapper.eq("username", userDTO.getUsername());
+        if (userDTO.getUserName()!=null&&userDTO.getUserName().length()>0){
+            queryWrapper.eq("user_name", userDTO.getUserName());
         }
-        if (userDTO.getPhone()!=null&&userDTO.getPhone().length()>0){
-            queryWrapper.eq("phone", userDTO.getPhone());
-        }
-        if (userDTO.getMail()!=null&&userDTO.getMail().length()>0){
-            queryWrapper.eq("mail", userDTO.getMail());
+        if (userDTO.getUserPhone()!=null&&userDTO.getUserPhone().length()>0){
+            queryWrapper.eq("user_phone", userDTO.getUserPhone());
         }
         User user = userMapper.selectOne(queryWrapper);
         UserVO userVO = new UserVO();
         BeanUtils.copyProperties(user,userVO);
         if (userVO.getId()!=null && userVO.getId() > 0){
-            result = new Result.Builder(Status.SUCCESS.getStatus(), Message.SUCCESS.getMsg()).data(userVO).build();
+            dataResult = DataResult.getBuilder(Status.SUCCESS.getStatus(), Message.SUCCESS.getMsg()).data(userVO).build();
         }else {
             return null;
         }
-        return result;
+        return dataResult;
     }
-
 
     /**
      * 添加用户
@@ -106,35 +105,38 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IU
      * @return
      */
     @Override
-    public Result addUser(UserDTO userDTO){
-        Result result;
+    public HttpResult addUser(UserDTO userDTO){
+        DataResult dataResult;
         // 设置默认用户组
-        userDTO.setUsergroup(GroupEnum.USER_GROUP.getGroupId());
+        userDTO.setUserGroup(GroupEnum.USER_GROUP.getGroupId());
         // 若未传入密码，则设置默认密码为123456
-        if (userDTO.getPassword()==null || userDTO.getPassword().length()==0){
-            userDTO.setPassword(passwordService.createPassword("123456"));
+        if (userDTO.getUserPassword()==null || userDTO.getUserPassword().length()==0){
+            userDTO.setUserPassword(passwordService.createPassword("123456"));
         }
-        if (userDTO.getUsername()==null || userDTO.getUsername().length()==0){
-            result = new Result.Builder(Status.INSERT_ERROR.getStatus(), Message.INSERT_ERROR.getMsg()).build();
-            return result;
+        if (userDTO.getUserName()==null || userDTO.getUserName().length()==0){
+            dataResult = DataResult.getBuilder(Status.INSERT_ERROR.getStatus(), Message.INSERT_ERROR.getMsg()).build();
+            return dataResult;
         }
-        if (userDTO.getPhone()==null || userDTO.getPhone().length()==0){
-            result = new Result.Builder(Status.INSERT_ERROR.getStatus(), Message.INSERT_ERROR.getMsg()).build();
-            return result;
+        if (userDTO.getUserPhone()==null || userDTO.getUserPhone().length()==0){
+            dataResult = DataResult.getBuilder(Status.INSERT_ERROR.getStatus(), Message.INSERT_ERROR.getMsg()).build();
+            return dataResult;
+        }
+        // 判断用户是否已经存在
+        if (userMapper.selectCount(
+                Wrappers.query(new User())
+                        .eq("user_name", userDTO.getUserName())
+                        .eq("user_phone", userDTO.getUserPhone())) > 0){
+            return DataResult.getBuilder(Status.USER_EXISTS.getStatus(), Message.USER_EXISTS.getMsg()).data(userDTO).build();
         }
         User user = new User();
-        UserVO userVO = new UserVO();
         BeanUtils.copyProperties(userDTO, user);
-        BeanUtils.copyProperties(userDTO, userVO);
         int addResult = userMapper.insert(user);
-        if (addResult>0){
-            return new Result.Builder(Status.SUCCESS.getStatus(), Message.SUCCESS.getMsg()).data(user).build();
-        }
-        else {
-            return new Result.Builder(Status.INSERT_ERROR.getStatus(), Message.INSERT_ERROR.getMsg()).build();
+        if (addResult > 0){
+            return selectOne(userDTO);
+        }else {
+            return DataResult.getBuilder(Status.INSERT_ERROR.getStatus(), Message.INSERT_ERROR.getMsg()).build();
         }
     }
-
 
     /**
      * 删除一个用户
@@ -142,28 +144,28 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IU
      * @return
      */
     @Override
-    public Result deleteOne(UserDTO userDTO){
-        Result result;
+    public HttpResult deleteOne(UserDTO userDTO){
+        DataResult dataResult;
         QueryWrapper<User> queryWrapper = new QueryWrapper<>();
         if (userDTO.getId()!=null && userDTO.getId()>0){
             queryWrapper.eq("id", userDTO.getId());
         }
-        if (userDTO.getUsername()!=null && userDTO.getUsername().length()>0){
-            queryWrapper.eq("username",userDTO.getUsername());
+        if (userDTO.getUserName()!=null && userDTO.getUserName().length()>0){
+            queryWrapper.eq("user_name",userDTO.getUserName());
         }
         List<User> list = userMapper.selectList(queryWrapper);
         if (list == null || list.size() != 1){
-            result = new Result.Builder(Status.QUERY_ERROR.getStatus(), Message.QUERY_ERROR.getMsg()).build();
+            dataResult = DataResult.getBuilder(Status.QUERY_ERROR.getStatus(), Message.QUERY_ERROR.getMsg()).build();
         }else {
             User user = list.get(0);
-            user.setDel(true);
+            user.setDelFlag(1);
             int update = userMapper.update(user, queryWrapper);
             if (update == 0){
-                result = new Result.Builder(Status.UPDATE_ERROR.getStatus(),Message.UPDATE_ERROR.getMsg()).build();
+                dataResult = DataResult.getBuilder(Status.UPDATE_ERROR.getStatus(),Message.UPDATE_ERROR.getMsg()).build();
             }
-            result = new Result.Builder().build();
+            dataResult = DataResult.getBuilder().data(UserTransformer.user2UserVo(user)).build();
         }
-        return result;
+        return dataResult;
     }
 
     /**
@@ -172,27 +174,23 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IU
      * @return
      */
     @Override
-    public Result exchange(UserDTO userDTO){
-        Result result;
-        User user = new User();
+    public HttpResult exchange(UserDTO userDTO){
+        DataResult dataResult;
+        if (Objects.isNull(userDTO.getId())){
+            return DataResult.getBuilder(Status.PARAM_NEED.getStatus(), Message.PARAM_NEED.getMsg()).build();
+        }
         // 密码加密
-        if (userDTO.getPassword()!=null && userDTO.getPassword().length()>0){
-            userDTO.setPassword(passwordService.createPassword(userDTO.getPassword()));
+        if (userDTO.getUserPassword()!=null && userDTO.getUserPassword().length()>0){
+            userDTO.setUserPassword(passwordService.createPassword(userDTO.getUserPassword()));
         }
-        BeanUtils.copyProperties(userDTO, user);
-        UpdateWrapper<User> userUpdateWrapper = new UpdateWrapper<>();
-        userUpdateWrapper.eq("id", userDTO.getId());
-
-        int update = userMapper.update(user, userUpdateWrapper);
-        UserVO userVO = new UserVO();
+        int update = userMapper.updateById(UserTransformer.userDTO2User(userDTO));
         if (update>0){
-            BeanUtils.copyProperties(userDTO, userVO);
-            result = new Result.Builder().data(userVO).build();
+            User completeUser = userMapper.selectById(userDTO.getId());
+            dataResult = DataResult.getBuilder().data(UserTransformer.user2UserVo(completeUser)).build();
+        }else {
+            dataResult = DataResult.getBuilder(Status.UPDATE_ERROR.getStatus(), Message.UPDATE_ERROR.getMsg()).data(userDTO).build();
         }
-        else {
-            result = new Result.Builder(Status.UPDATE_ERROR.getStatus(), Message.UPDATE_ERROR.getMsg()).data(userDTO).build();
-        }
-        return result;
+        return dataResult;
     }
 
     /**
@@ -201,24 +199,22 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IU
      * @return
      */
     @Override
-    public Result login(UserDTO userDTO) {
-        String queryConditionName = "username";
+    public HttpResult login(UserDTO userDTO) {
+        String queryConditionName = "user_name";
         QueryWrapper<User> queryWrapper = new QueryWrapper<>();
-        if (null == userMapper.selectOne(queryWrapper.eq(queryConditionName, userDTO.getUsername()))){
-            return new Result.Builder(Status.USER_NOT_FOUND.getStatus(), Message.USER_NOT_FOUND.getMsg()).build();
+        if (null == userMapper.selectOne(queryWrapper.eq(queryConditionName, userDTO.getUserName()))){
+            return DataResult.getBuilder(Status.USER_NOT_FOUND.getStatus(), Message.USER_NOT_FOUND.getMsg()).build();
         }
-        User user = userMapper.selectOne(queryWrapper.eq(queryConditionName, userDTO.getUsername()));
-        if (!passwordService.authenticatePassword(user.getPassword(), userDTO.getPassword())){
-            return new Result.Builder(Status.ERROR_PASSWORD.getStatus(), Message.ERROR_PASSWORD.getMsg()).build();
+        User user = userMapper.selectOne(queryWrapper.eq(queryConditionName, userDTO.getUserName()));
+        if (!passwordService.authenticatePassword(user.getUserPassword(), userDTO.getUserPassword())){
+            return DataResult.getBuilder(Status.ERROR_PASSWORD.getStatus(), Message.ERROR_PASSWORD.getMsg()).build();
         }
         // 验证通过，将当前用户存入session中
-        Person p = new Person();
-        PersonVo pv = new PersonVo();
-        BeanUtils.copyProperties(user, pv);
-        BeanUtils.copyProperties(pv, p);
+        Person p = UserTransformer.user2Person(user);
         UserContext.putCurrentUser(p);
-
-        return new Result.Builder(Status.SUCCESS.getStatus(), Message.WELCOME_TO_LOGIN.getMsg()).data(p).build();
+        return DataResult.getBuilder(Status.SUCCESS.getStatus(), Message.WELCOME_TO_LOGIN.getMsg())
+                .data(UserTransformer.user2UserVo(user))
+                .build();
     }
 
     /**
@@ -227,10 +223,18 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IU
      * @return
      */
     @Override
-    public Result logout() {
+    public HttpResult logout() {
+        PersonVo currentUser = UserContext.getCurrentUser();
+        if (Objects.isNull(currentUser)){
+            return DataResult.fail().build();
+        }
+        UserVO userVO = UserTransformer.personVO2UserVO(currentUser);
         // 验证通过，将当前用户从session中删除
         UserContext.deleteCurrentUser();
-        return new Result.Builder(Status.SUCCESS.getStatus(), Message.BYE_BYE.getMsg()).build();
+        return new DataResult
+                .Builder(Status.SUCCESS.getStatus(), Message.BYE_BYE.getMsg())
+                .data(userVO)
+                .build();
     }
 
     /**
@@ -238,13 +242,13 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IU
      * @return
      */
     @Override
-    public Result getCurrentUser(){
+    public HttpResult getCurrentUser(){
         PersonVo person = new PersonVo();
         try {
             person = UserContext.getCurrentUser();
         }catch (IllegalArgumentException e){
             e.fillInStackTrace();
         }
-        return new Result.Builder().data(person).build();
+        return DataResult.getBuilder().data(person).build();
     }
 }
